pyproject.toml
[project]
name = "config-guidance-backend"
version = "0.1.0"
description = "Backend for Configuration Guidance System"
dependencies = [
    "pdfplumber>=0.10.2",
    "pyyaml>=6.0.1",
    "beautifulsoup4>=4.12.2",
    "html2text>=2020.1.16",
    "sentence-transformers>=2.2.2",
    "langchain>=0.0.300",
    "langchain-community>=0.0.300",
    "faiss-cpu>=1.7.4",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.2"
]

[project.scripts]
start = "backend.main:app"